+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/api_client/__init__.py
@@ -7,9 +7,9 @@ from ._public import (
 )
 
 __all__: tuple[str, ...] = (
-    "DynamicSidecarClient",
-    "setup_api_client",
     "close_api_client",
+    "DynamicSidecarClient",
     "get_dynamic_sidecar_client",
+    "setup_api_client",
     "update_dynamic_sidecar_health",
 )
