+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/api_client/_public.py
@@ -18,7 +18,7 @@ from ..errors import (
     NodeportsDidNotFindNodeError,
 )
 from ._errors import UnexpectedStatusError
-from ._thin_client import ThinDynamicSidecarClient
+from ._thin import ThinDynamicSidecarClient
 
 logger = logging.getLogger(__name__)
